docker_api_exporter/requirements.txt
prometheus_client
docker
pyyaml